requirements.txt
# Core
torch>=2.0
transformers>=4.35
sentence-transformers>=2.2

# LLM APIs
openai>=1.0
anthropic>=0.18

# Environments
gymnasium>=0.29
# webshop - install from source
# alfworld - install from source

# Data
jsonlines>=4.0
pandas>=2.0
numpy>=1.24

# Training
wandb>=0.16
tqdm>=4.65
pyyaml>=6.0

# Analysis
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
umap-learn>=0.5

# Dev
pytest>=7.0
black>=23.0
ipykernel>=6.0
